Rename component2 and coupled2 to component and coupled
