+++ macros/src/component/atomic.rs
