+++ macros/src/component/backend.rs
@@ -22,7 +22,7 @@ use syn::{
     Error, Ident, LitInt, Result, Token,
 };
 
-use crate::component2::CommonComponent;
+use crate::component::CommonComponent;
 
 /// Generated token fragments used to construct backend channel code.
 pub struct ChannelTokens {
