+++ macros/src/component/backend/embassy.rs
@@ -1,5 +1,5 @@
 use super::{Backend, ChannelTokens, RtEngineArgs};
-use crate::component2::CommonComponent;
+use crate::component::CommonComponent;
 use heck::ToShoutySnakeCase;
 use syn::{
     parse::{Parse, ParseStream},
