+++ macros/src/component/backend/tokio.rs
@@ -1,5 +1,5 @@
 use super::{Backend, ChannelTokens, RtEngineArgs};
-use crate::component2::CommonComponent;
+use crate::component::CommonComponent;
 use proc_macro2::{Span, TokenStream as TokenStream2};
 use syn::{
     parse::{Parse, ParseStream},
