+++ macros/src/component/coupled.rs
