+++ macros/src/component/state.rs
